backEnd/requirements.txt
amqp==2.5.2
anyjson==0.3.3
appdirs==1.4.3
attrs==19.3.0
Automat==0.8.0
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
dbus-python==1.2.16
distlib==0.3.0
distro==1.4.0
Django==2.2.12
djangorestframework==3.12.4
entrypoints==0.3
ephem==3.7.7.1
filelock==3.0.12
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
Jinja2==2.10.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
kombu==4.6.7
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
MarkupSafe==1.1.0
more-itertools==4.2.0
nested-lookup==0.2.22
netifaces==0.10.4
oauthlib==3.1.0
pexpect==4.6.0
psycopg2-binary==2.9.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pyrsistent==0.15.5
pyserial==3.4
python-dateutil==2.7.3
python-decouple==3.5
python-memcached==1.58
pytz==2019.3
PyYAML==5.3.1
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
sentry-sdk==1.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sqlparse==0.2.4
ssh-import-id==5.10
Twisted==18.9.0
urllib3==1.25.8
vine==1.3.0
virtualenv==20.0.17
wadllib==1.3.3
zipp==1.0.0
zope.interface==4.7.1
gunicorn
whitenoise
